SVM.jl

SVM.jl

This unit implements the Support Vector models for the data points in the manifold of positive definite matrices. The structures below are similar to classes of LinearSVC and SVC we have in ScikitLearn in Python. This Module incorporates two models of Support Vector Machine type : - LinearSVM - SVM

For further information on LinearSVM and SVM one may refer to the scikit-learn documentations of LinearSVC and SVC respectively.

An object of scikit-learn LinearSVC or SVC is automatically created by the constructor as soon as an instance of their structs is created by the user.

LinearSVM structure

RiemannianML.LinearSVM — Type.

Structure for implementing LinearSVC, with all the arguments as attributes to this structure. The 'clf' attribute is the python object of LinearSVC that is constructed by the constructor function of this structure as soon as a LinearSVM instance is created by the user.

Parameters :

penalty:: String

loss :: String

dual :: Bool

tol :: Float64

C :: Float64

multi_class :: String

fit_intercept :: Bool

intercept_scaling :: Float64

class_weight :: Dict

verbose :: Int64

random_state :: Int

max_iter :: Int64

One should refer to the ScikitLearn documentation of LinearSVC for the better understanding of these parameters with respect to Linear Support Vector Machine.

## Example
using RiemannianML
model = LinearSVM()

source

SVM structure

RiemannianML.SVM — Type.

Structure for implementing SVC, with all the arguments as attributes to this structure. The 'clf' attribute is the python object of SVC that is constructed by the constructor function of this structure as soon as a SVM instance is created by the user.

Parameters :

penalty:: String

loss :: String

dual :: Bool

tol :: Float64

C :: Float64

multi_class :: String

fit_intercept :: Bool

intercept_scaling :: Float64

class_weight :: Dict

verbose :: Int64

random_state :: Int

max_iter :: Int64

One should refer to the ScikitLearn documentation of SVC for the better understanding of these parameters with respect to Support Vector Machine.

## Example
using RiemannianML
model = SVM()

source