cross_mdm.jl

cross_mdm.jl

This unit contains cross validation algorithm for MDM( Minimum Distance to Mean) classifier i.e. applied directly on Positive Definite Manifold. Similar to ScikitLearn in Python, one can have a better evaluation of the classifier by using cross validation. Unit contains supporting function indCV and function crossvalmdm.

Function	Description

indCV	returns the vectors containing shuffled indices of training and testing samples for each cross validation iteration

cross_val_mdm	implements cross validation for MDM classifier

indCV

RiemannianML.indCV — Function.

This is a helper function to implement crossvalmdm. It returns the vectors containing indices of training and testing samples for each cross validation iteration. The indices so received are shuffled well and not contiguous strings or blocks of data samples. This function uses shuffle! to shuffle all the indices and then divides the shuffled set into training and testing sets. This function holds the prime basis of CrossValidation implementation.

Arguments taken by this function are:

k::Int :- Last number of the sequence of natural numbers to be shuffled starting from 1

nCV:: Int :- Number of cross-validation for which the indices are to be generated.

shuffle :- Boolean to inform whether to do shuffling or not. Default is set to True.

Return values :

nTest :- The size of each testing set.

nTrain :- The size of each training set.

indTrain :- The list of all the vectors that contain the training indices for each iteration.

indTest :- The list of all the vectors that contain the testing indices for each iteration.

crossvalmdm

RiemannianML.cross_val_mdm — Function.

This function performs cv number of iterations in each of which it runs the mdm model over a part of the shuffled data. This way helps us to have a better evaluation of our classifier.

This is the main function implementing cross validation for MDM classifier. It uses indCV as its basic helper function. It returns the vector of confusion matrices, coreesponding to each cross valdation iteration and also the final confusion matrix i.e. the sum of all the previous ones. The function also returns average balanced accuracy or average regular accuracy depending on the choice of the user. The default is set to balanced accuracy type. It takes the following arguments:

Arguments taken by this function are:

𝐗 :: ℍVector :- The training set of type Vector of Hermitian matrices.

y :- The list of labels corresponding to each trial

cv :: Int :- The number of cross-validation desired by the user

scoring :: String :- If Balanced Accuracy is requied or the Regular Accuracy. The default is set to Balanced Accuracy.

metric :: Metric :- The metric space in which to do the computations. To be specified by the user as one of the many metric spaces options provided in PostDefManifold. One may refer to [mdm.jl documentations] for exploring all the possible options. The default is set to Fisher.

Return value:

cross_val score :- The list of the balanced accuracy score for each cross_val iteration.