logisticRegression.jl

logisticRegression.jl

This unit implements the LogisticRegression models for the data points in the Riemannian manifold of Symmetric Positive Definite (SPD). The structures below are similar to classes of LogisticRegression and LogisticRegressionCV we have in ScikitLearn in Python. This Module incorporates two models of LogisticRegression : - LogisticReg - Simple logisticRegression in which the value of the penalty coefficient alpha is set manually by the user. The user is responsible for choosing a specific value of alpha. - LogisticRegCV - LogisticRegression with cross validation, to find the best suit for alpha. The user can provide a range of alpha and the algorithm finds the best suited value of alpha out of those.

For further information on LogisticReg and LogisticRegCV one may refer to the scikit-learn documentations of LogisticRegression and LogisticRegressionCV respectively.

An object of scikit-learn LogisticRegression or LogisticRegressionCV is automatically created by the constructor as soon as an instance of their structs is created by the user.

LogisticReg structure

RiemannianML.LogisticReg — Type.

Structure for implementing LogisticRegression, with all the arguments as attributes to this structure. The 'clf' attribute is the python object of LogisticRegression that is constructed by the constructor function of this structure as soon as a LogisticReg instance is created by the user.

Parameters :

penalty:: String

dual :: Bool

tol :: Float64

C :: Float64

fit_intercept :: Bool

intercept_scaling :: Float64

class_weight :: Dict

random_state :: Int

solver :: String

max_iter :: Int64

multi_class :: String

verbose :: Int64

warm_start :: Bool

n_jobs :: Int

l1_ratio :: Float64

One may refer to the ScikitLearn documentation of LogisticRegression for the better understanding of these parameters with respect to LogisticRegression.

## Example
using RiemannianML
model = LogisticReg()

source

LogisticRegCV structure

RiemannianML.LogisticRegCV — Type.

Structure for implementing LogisticRegressionCV, with all the arguments as attributes to this structure. The 'clf' attribute is the python object of LogisticRegressionCV that is constructed by the constructor function of this structure as soon as a LogisticRegCV instance is created by the user.

Parameters :

Cs :: Int[] or Float64[]

fit_intercept :: Bool

cv :: Int

dual :: Bool

penalty :: String

scoring :: String

solver :: String

tol :: Float64

max_iter :: Int

class_weight :: Dict

n_jobs :: Int

verbose :: Int

refit :: Bool

intercept_scaling :: Float64

multi_class :: String

random_state :: Int

l1_ratios :: Float64

One should refer to the ScikitLearn documentation of LogisticRegressionCV for the better understanding of these parameters with respect to LogisticRegressionCV.

## Example
using RiemannianML
model = LogisticRegCV()

source