mdm.jl

mdm.jl

This unit implemets the MDM (Minimum Distance to Mean) classifier for the manifold of positive definite matrices. Similarly to what is done in ScikitLearn in Python, a type is created (struct in Julia) of the desired specifications. It also implements cross validation algorithm for MDM( Minimum Distance to Mean) classifier again similar to ScikitLearn in Python, one can have a better evaluation of the classifier by using cross validation. Module incorporates supporting functions : find_dist, predict_mdm, predict_prob, indCV.

It implemens a structure MDM and includes the following functions :

Function	Description

mean_mdm	calculates the mean of all the classes in the training set and also notifies the user if the mean is not convergent in case of some metric spaces

find_dist	finds the distance of each sample case from the so found means of all the classes

predict_mdm	predicts the class for each sample case depending on its distance from the respective means

predict_prob	predicts the probability of each class for each sample case depending on its distance from the respective means

indCV	returns the vectors containing shuffled indices of training and testing samples for each cross validation iteration

cross_val_mdm	implements cross validation for MDM classifier

For a detailed understanding of mdm, one should know the basics of Riemannian Geometry and its application in the classification of positive definite matrices. One may refer to the following papers for getting the feel of the process.

A. Barachant, S. Bonnet, M. Congedo, C. Jutten (2012)🎓

A. Barachant, S. Bonnet, M. Congedo, C. Jutten (2013)🎓

M. Congedo, A. Barachant, R. Bhatia R (2017a)🎓

M. Congedo, A. Barachant, E. Kharati Koopaei (2017b)🎓

Or one may directly look into the Intro to Riemannian Geometry section of the PostDefManifold documentation and quench all their doubts.

MDM structure

RiemannianML.MDM — Type.

This is a structure of the MDM model. It has two attributes namely metric and class_means.

metric :: Metric :- The user needs to specify the metric space in which all the distance computations is to be done. The possible options are:

Metric	Mean estimation

Euclidean	distance: δ_e; mean: Arithmetic

invEuclidean	distance: δ_i; mean: Harmonic

ChoEuclidean	distance: δ_c; mean: Cholesky Euclidean

logEuclidean	distance: δ_l; mean: Log Euclidean

logCholesky	distance: δ_c; mean: Log-Cholesky

Fisher	distance: δ_f; mean: Fisher (Cartan, Karcher, Pusz-Woronowicz,...)

logdet0	distance: δ_s; mean: LogDet (S, α, Bhattacharyya, Jensen,...)

Jeffrey	distance: δ_j; mean: Jeffrey (symmetrized Kullback-Leibler)

VonNeumann	distance: δ_v; mean: Not Availale

Wasserstein	distance: δ_w; mean: Wasserstein (Bures, Hellinger, ...)

class_means :- This is not to be specified by the user. This comes to play when a model is fit with a set of training data so we have to store the means corresponding to each of the classes. We can directly access this for an already fit model while using functions like predict.

source

mean_mdm

RiemannianML.mean_mdm — Function.

Given a metric, ℍVector( Vector of Hermitian matrices), weights(optional) returns the mean of these Hermitian matrices for mdm classifier.This function is kind of an interface to the mean function of PostDefManifold. Refer to the Mean of PosDefManifold.

Arguments :

metric :: Metric :-The user needs to specify the metric space in which all the distance computations is to be done.

𝐏::ℍVector :-Vector of Hermitian matrices or simply a HermitianVector. The vector of points in the Symmetric Positive Definite manifold to be transformed into the the tangent space.

w::Vector(optional):- Vector containing weights corresponding to every point in 𝐏.

✓w = true(optional) :- Boolean to determine whether to calculate weighted mean or just take w = []. It also checks if the weights sum up to 1. If not does normalization.

⏩ = false (optional) :- Boolean to allow threading or not.

Returns :

G :: ℍ :- Mean of the set of ℍVector.

source

find_dist

RiemannianML.find_dist — Function.

Given a sample set, classmeans, type of metric returns the distance of each sample case from the points in classmeans. Distance should be caluclated in the metric space opted by the user while creating the instance of mdm.

Arguments :

sample :: ℍVector :-The vector of Hermitian matrices or points in the positive definite manifold for which the prediction is to be made using the model(argument 1) already been trained specially for it.

class_means :- Vector of Hermitian matrices that represent the class means or the centroid of all the classes in the training set.

metric :: Metric :-The user needs to specify the metric space in which all the distance computations is to be done.

Return :

A :: Array{Float64,2} :- Array of distances of each of the sample case from each of the class_means.

source

predict_mdm

RiemannianML.predict_mdm — Function.

Given a sample set, class_means, type of metric returns the predicted classes for each sample case according to the Minimum Distance to the Means scheme.

Arguments :

sample :: ℍVector :- The vector of Hermitian matrices or points in the positive definite manifold for which the prediction is to be made using the model(argument 1) already been trained specially for it.

class_means :- Vector of Hermitian matrices that represent the class means or the centroid of all the classes in the training set.

metric :: Metric :- The user needs to specify the metric space in which all the distance computations is to be done.

Returns :

result :: Array{Int, 1} :- The List of the predicted classes for the given sample set.

source

predict_prob

RiemannianML.predict_prob — Function.

Given a sample set, class_means, type of metric returns the probability of each class for each sample case according to the Minimum Distance to the Means scheme. All the probability sums to 1. This function makes use of the softmax function from thr PostDefManifold to calculate the probability values.

Arguments :

sample :: ℍVector :-The vector of Hermitian matrices or points in the positive definite manifold for which the probability prediction is to be made using the model(argument 1) already been trained specially for it.

class_means :- Vector of Hermitian matrices that represent the class means or the centroid of all the classes in the training set.

metric :: Metric :-The user needs to specify the metric space in which all the distance computations is to be done.

Returns :

Prob :: Array{Float64,1}:- The List of the predicted probabilities corresponding to each of the classes for the given sample set.

source

indCV

RiemannianML.indCV — Function.

Given the length, nCV( number of cross-validations) returns the vectors containing indices of training and testing samples for each cross validation iteration. This is a helper function to implement crossvalmdm. It uses shuffle! of the Random.jl package. This function holds the prime basis of CrossValidation implementation.

Arguments :

k::Int :- Last number of the sequence of natural numbers to be shuffled starting from 1

nCV:: Int :- Number of cross-validation for which the indices are to be generated.

shuffle :- Boolean to inform whether to do shuffling or not. Default is set to True.

Returns :

nTest :- The size of each testing set.

nTrain :- The size of each training set.

indTrain :- The list of all the vectors that contain the training indices for each iteration.

indTest :- The list of all the vectors that contain the testing indices for each iteration.

source

crossvalmdm

RiemannianML.cross_val_mdm — Function.

Given a ℍVector(training set), y(labels), cv(number of cross-validations) returns the the list containing the score for each cross-validation iteration. The score may be average balanced accuracy or average regular accuracy depending on the choice of the user. *It is better to call for balanced accuracy as a better estimation of the model performance. Overall, if the number of ases in for each class are same, balanced accuracy equals the regular accuracy.

This is the main function implementing cross validation for MDM classifier. It uses indCV as its basic helper function. It also returns the final confusion matrix.

Arguments :

𝐗 :: ℍVector :- The training set of type Vector of Hermitian matrices.

y :- The list of labels corresponding to each trial

cv :: Int :- The number of cross-validation desired by the user

scoring :: String :- If Balanced Accuracy is requied or the Regular Accuracy. The default is set to Balanced Accuracy.

metric :: Metric :- The metric space in which to do the computations. To be specified by the user as one of the many metric spaces options provided in PostDefManifold. One may refer to [mdm.jl documentations] for exploring all the possible options. The default is set to Fisher.

cnfmat :: Bool :- Boolean to notify if the user wants the confusion matrix also. Default is set to false.

Returns :

cross_val score :- The list of the balanced accuracy score for each cross_val iteration.

source