knn.jl

knn.jl

This unit implements the KNeighborsClassifier model for the data points in the Riemannian manifold of Symmetric Positive Definite (SPD). The structure below is similar to a class of KNeighborsClassifier we have in ScikitLearn in Python. For further information on kneighborClf, one may refer to the scikit-learn documentation of KNeighborsClassifier.

An object of scikit-learn KNeighborsClassifier is automatically created by the constructor as soon as an object of this struct is created by the user.

kneighborClf structure

RiemannianML.kneighborClf — Type.

Structure for implementing KNeighborsClassifier. The clf attribute is the python object of KNeighborsClassifier that is constructed by the constructor function of this structure as soon as a kneighborClf instance is created by the user.

Parameters :

n_neighbors :: Int64

weights:: String

algorithm:: String

leaf_size:: Int64

p:: Int64

metric:: String

metric_params :: Dict

n_jobs :: Int

One may refer to the ScikitLearn documentation of KNeighborsClassifier for the better understanding of these parameters with respect to kneighborClf.

## Example
using RiemannianML
model = kneighborClf(n_neighbors=3)

source